+++ kernel/arch/x86/drivers/usb_xhci_event.c
@@ -129,7 +129,7 @@ bool _xhci_process_events(
                     ctrl->bus,
                     ctrl->slot,
                     ctrl->func,
-                    type
+                    (unsigned int)type
                 );
                 continue;
             }
